==> Step_1-Match_Results.ipynb <==
import pandas as pd

import numpy as np

## 1. Data Loadind

#### Raw data parsed from https://www.resultados.com/futebol/brasil/serie-a/, transformed into a spreadsheet and expoted to a CSV file.

jogos = pd.read_csv("data/brasileiro-a.9.csv")

jogos.head()

jogos.info()

## 2. Data Parsing 

### 2.1 Remove white spaces

# select columns of type object
jogos_obj = jogos.select_dtypes(['object'])

jogos[jogos_obj.columns] = jogos_obj.apply(lambda x: x.str.strip())

### 2.2 List soccer teams from all seasons

#### 2.2.1 Get unique mandatory teams 

times = list(jogos["mandante"].unique())

#### 2.2.2 Sort teams by name

times.sort()

#### 2.2.3 Check the teams

times

#### 2.2.4 Check the number of times

len(times)

### 2.3 Transform score in two columns

#### 2.3.1 Defines a function to parse the score

# Function returns first part of score if is the home team or second part otherwise
def tratarPlacar(row, mandante):
    placar = row["placar"].split(" : ")
    if (mandante):
        return int(placar[0])
    return int(placar[1])

#### 2.3.2 Fills the goals of the home team in to a column

jogos['gols_mandante'] = jogos.apply (lambda row: tratarPlacar(row, True),axis=1)

#### 2.3.2 Fills the goals of the visitor team in to a column

jogos['gols_visitante'] = jogos.apply (lambda row: tratarPlacar(row, False),axis=1)

#### 2.3.3 Remove scrore column

del jogos["placar"]

#### 2.3.4 Describe the matches DataFrame

jogos.describe()

#### 2.3.5 Check the matches DataFrame

jogos.head(2)

### 2.4 Parse the date column

#### 2.4.1 Defines a function to parse the date

# Converts the date from dd.mm to dd-mm format
def tratarData(row):
    data = row["data"].split(".")    
    return data[0] + "-" + data[1]

#### 2.4.2 Parse data form "data" column

jogos['data'] = jogos.apply (lambda row: tratarData(row),axis=1)

#### 2.4.3 Check the result on  "data"column 

jogos.head(2)

### 2.5 Create a result column 

#### 2.5.1 Defines a function to parse score

# Retuns "MANDANTE" if the home team wins the macth,
# "VISITANTE" if visitor wins or "EMPATE" in case of tied match
def resultado(row):
    if (row["gols_mandante"] > row["gols_visitante"]):
        return "MANDANTE"
    elif (row["gols_mandante"] < row["gols_visitante"]):
        return "VISITANTE"
    else:
        return "EMPATE"

#### 2.5.2 Apply the function to "resultado" column

jogos['resultado'] = jogos.apply (lambda row: resultado(row),axis=1)

#### 2.5.3 Check the matches DataFrame

jogos.head(2)

### 2.6 Export matches DataFrame to file

jogos.to_csv("data/match-results.9.csv", sep=",")
