==> Step_2-Team_Stars.ipynb <==
import pandas as pd

import numpy as np

## 1. Data Loadind

jogos = pd.read_csv("data/match-results.7.csv")

del jogos["Unnamed: 0"]

jogos.head()

jogos.info()

times = list(jogos["mandante"].unique())

## 2. Checking results 

### 2.1 Results grouped

totais_resultados = jogos.groupby("resultado").size()

### 2.2 Total  of matches

total_jogos = totais_resultados.sum()

### 2.3 Create a DataFrame with grouped results

resultados_DF = pd.DataFrame({'resultados': totais_resultados.keys(), 'totais': totais_resultados.values})

resultados_DF

def percentual(valor, total):
    return round(valor/total, 4)

resultados_DF['%'] = resultados_DF.apply (lambda row: percentual(row["totais"], total_jogos),axis=1)

resultados_DF

### Calcular resultados de um time

def resultadosTime(time, casa, resultado):
    return len(jogos[(jogos[casa] == time) & (jogos["resultado"] == resultado)]) 

resultados_time_DF = pd.DataFrame({'time': times })

resultados_time_DF["vitorias_casa"] = resultados_time_DF.apply (lambda row: resultadosTime(row["time"], "mandante", "MANDANTE"),axis=1)
resultados_time_DF["empates_casa"] = resultados_time_DF.apply (lambda row: resultadosTime(row["time"], "mandante", "EMPATE"),axis=1)
resultados_time_DF["derrotas_casa"] = resultados_time_DF.apply (lambda row: resultadosTime(row["time"], "mandante", "VISITANTE"),axis=1)
resultados_time_DF["vitorias_fora"] = resultados_time_DF.apply (lambda row: resultadosTime(row["time"], "visitante", "VISITANTE"),axis=1)
resultados_time_DF["empates_fora"] = resultados_time_DF.apply (lambda row: resultadosTime(row["time"], "visitante", "EMPATE"),axis=1)
resultados_time_DF["derrotas_fora"] = resultados_time_DF.apply (lambda row: resultadosTime(row["time"], "visitante", "MANDANTE"),axis=1)

resultados_time_DF.head(3)

resultados_time_DF.eval("jogos_casa = vitorias_casa + empates_casa + derrotas_casa", inplace=True)
resultados_time_DF.eval("jogos_fora = vitorias_fora + empates_fora + derrotas_fora", inplace=True)
resultados_time_DF.eval("jogos_total = jogos_fora + jogos_fora", inplace=True)

resultados_time_DF.head(3)

def aproveitamentoMandante(row):
    total_jogos_mandante = row["vitorias_casa"] + row["empates_casa"] + row["derrotas_casa"]
    return (3*row["vitorias_casa"] + row["empates_casa"]) / (3*total_jogos_mandante)

resultados_time_DF["aprov_casa"] = resultados_time_DF.apply (lambda row: aproveitamentoMandante(row),axis=1)
resultados_time_DF.head(3)


def aproveitamentoVisitante(row):
    total_jogos_visitante = row["vitorias_fora"] + row["empates_fora"] + row["derrotas_fora"]
    return (3*row["vitorias_fora"] + row["empates_fora"]) / (3*total_jogos_visitante)

resultados_time_DF["aprov_fora"] = resultados_time_DF.apply (lambda row: aproveitamentoVisitante(row),axis=1)
resultados_time_DF.head(3)

def aproveitamentoTotal(row):
    total_jogos_mandante = row["vitorias_casa"] + row["empates_casa"] + row["derrotas_casa"]
    total_jogos_visitante = row["vitorias_fora"] + row["empates_fora"] + row["derrotas_fora"]
    return (3*row["vitorias_fora"] + row["empates_fora"] + 3*row["vitorias_casa"] + row["empates_casa"]) / (3*total_jogos_mandante + 3*total_jogos_visitante)

resultados_time_DF["aprov_total"] = resultados_time_DF.apply (lambda row: aproveitamentoTotal(row),axis=1)
resultados_time_DF.sort_values(["aprov_total"], ascending=[0]).head(3)

resultados_time_DF = resultados_time_DF[["time", "jogos_casa", "jogos_fora", "jogos_total", "aprov_casa", "aprov_fora", "aprov_total"]].sort_values(["aprov_total"], ascending=[0])

def golsMandante(time):
    return jogos[jogos["mandante"] == time][["gols_mandante"]].sum()

def golsVisitante(time):
    return jogos[jogos["visitante"] == time]["gols_visitante"].sum()

resultados_time_DF["gols_mandante"] = resultados_time_DF.apply (lambda row: golsMandante(row["time"]),axis=1)
resultados_time_DF["gols_visitante"] = resultados_time_DF.apply (lambda row: golsVisitante(row["time"]),axis=1)

resultados_time_DF.eval("total_gols = gols_mandante + gols_visitante", inplace=True)
resultados_time_DF.eval("media_gols_casa = gols_mandante/jogos_casa", inplace=True)
resultados_time_DF.eval("media_gols_fora = gols_visitante/jogos_fora", inplace=True)
resultados_time_DF.eval("media_gols = (gols_mandante + gols_visitante)/jogos_total", inplace=True)

resultados_time_DF.head()


resultados_time_DF = resultados_time_DF[["time", "jogos_total", "aprov_casa", "aprov_fora", "aprov_total", "media_gols_casa", "media_gols_fora", "media_gols"]]

resultados_time_DF.set_index("time", inplace=True)

resultados_time_DF.sort_values(["aprov_total"], ascending = [0]).head(3)

resultados_time_DF_DESC = resultados_time_DF.describe(percentiles=[0, .1, .2, .3, .4, .5, .6, .7, .8, .9, 1])

resultados_time_DF_DESC

percentis = resultados_time_DF_DESC.loc[["10%","20%","30%","40%","50%","60%","70%","80%","90%"]]

percentis

def get_stars(array,value):
    print ()
    idx = (np.abs(array-value)).argmin()
    near = array[idx][0]
    index = np.where(array==near)[0][0]
    
    if (value > near):
        index = index + 1  
       
    return index + 1

stars_DF = resultados_time_DF.copy()

stars_DF["aprov_total"] = resultados_time_DF.apply (lambda row: get_stars(percentis[["aprov_total"]].values, row["aprov_total"]),axis=1)

stars_DF["aprov_casa"] = resultados_time_DF.apply (lambda row: get_stars(percentis[["aprov_casa"]].values, row["aprov_casa"]),axis=1)

stars_DF["aprov_fora"] = resultados_time_DF.apply (lambda row: get_stars(percentis[["aprov_fora"]].values, row["aprov_fora"]),axis=1)

stars_DF["media_gols_casa"] = resultados_time_DF.apply (lambda row: get_stars(percentis[["media_gols_casa"]].values, row["media_gols_casa"]),axis=1)

stars_DF["media_gols_fora"] = resultados_time_DF.apply (lambda row: get_stars(percentis[["media_gols_fora"]].values, row["media_gols_fora"]),axis=1)

stars_DF["media_gols"] = resultados_time_DF.apply (lambda row: get_stars(percentis[["media_gols"]].values, row["media_gols"]),axis=1)

stars_DF.eval("forca_casa = (3*aprov_casa + media_gols_casa)/4", inplace=True)

stars_DF.eval("forca_fora = (3*aprov_fora + media_gols_fora)/4", inplace=True)

stars_DF.eval("forca = (3*aprov_total + 1*media_gols)/4", inplace=True)

stars_DF.sort_values(["forca"], ascending=[0])

stars_DF.to_csv(path_or_buf="data/team_stars.8.csv",sep=";")
