==> Step_3-Data_Analysis.ipynb <==
import pandas as pd

import numpy as np

## 1. Data Loadind

matches = pd.read_csv("data/match-results.9.csv")

matches.head(2)

matches.info()

stars = pd.read_csv("data/team_stars.csv", ';')

stars.head(2)

stars.info()

matches.groupby(["ano", "resultado"])[["resultado"]].count()

matches.groupby(["resultado"])[["ano"]].count()

### Totais por resultado

totais_resultados = matches.groupby("resultado").size()

total_jogos = totais_resultados.sum()

resultados_DF = pd.DataFrame({'resultados': totais_resultados.keys(), 'totais': totais_resultados.values})

resultados_DF

def percentual(valor, total):
    return round(valor/total, 4)

resultados_DF['%'] = resultados_DF.apply (lambda row: percentual(row["totais"], total_jogos),axis=1)

resultados_DF

seasons = list(jogos["ano"].unique())

seasons

def classificar(jogos, rodada, classificacao):
    
    classificacao_nova = classificacao.copy()
    
    for jogo in jogos[jogos["rodada"] == rodada].iterrows():
        jogo_info = jogo[1];
        if (jogo_info["resultado"] == "MANDANTE"):            
            classificacao_nova.loc[jogo_info["mandante"]]["pts"] = classificacao.loc[jogo_info["mandante"]]["pts"] + 3
            classificacao_nova.loc[jogo_info["mandante"]]["v"] = classificacao.loc[jogo_info["mandante"]]["v"] + 1            
        elif (jogo_info["resultado"] == "VISITANTE"):
            classificacao_nova.loc[jogo_info["visitante"]]["pts"] = classificacao.loc[jogo_info["visitante"]]["pts"] + 3
            classificacao_nova.loc[jogo_info["visitante"]]["v"] = classificacao.loc[jogo_info["visitante"]]["v"] + 1
        elif (jogo_info["resultado"] == "EMPATE"):     
            classificacao_nova.loc[jogo_info["mandante"]]["pts"] = classificacao.loc[jogo_info["mandante"]]["pts"] + 1
            classificacao_nova.loc[jogo_info["visitante"]]["pts"] = classificacao.loc[jogo_info["visitante"]]["pts"] + 1
            
        classificacao_nova.loc[jogo_info["mandante"]]["gp"] = classificacao.loc[jogo_info["mandante"]]["gp"] + jogo_info["gols_mandante"]    
        classificacao_nova.loc[jogo_info["mandante"]]["gc"] = classificacao.loc[jogo_info["mandante"]]["gc"] + jogo_info["gols_visitante"]
        
        classificacao_nova.loc[jogo_info["visitante"]]["gp"] = classificacao.loc[jogo_info["visitante"]]["gp"] + jogo_info["gols_visitante"]    
        classificacao_nova.loc[jogo_info["visitante"]]["gc"] = classificacao.loc[jogo_info["visitante"]]["gc"] + jogo_info["gols_mandante"]    
        
        classificacao_nova.loc[jogo_info["mandante"]]["jogos"] = classificacao.loc[jogo_info["mandante"]]["jogos"] + 1    
        classificacao_nova.loc[jogo_info["visitante"]]["jogos"] = classificacao.loc[jogo_info["visitante"]]["jogos"] + 1
    
    classificacao_nova.eval("sg = gp - gc", inplace=True)
    classificacao_nova.sort_values(["pts", "v", "sg", "gp"], ascending=[0,0,0,0], inplace=True)
        
    for i in range(1, len(classificacao_nova.index.values.tolist())+1):   
        classificacao_nova.loc[classificacao_nova.index[i-1]]["posicao"] = i
    
    return classificacao_nova

ranking_DF = {}

for season in seasons:         
    
    turns = list(matches.query("ano == " + str(season)).sort_values("rodada")["rodada"].unique())    
    teams = list(matches.query("ano == " + str(season))["mandante"].unique())
 
    ranking_DF[season] = {}
    
    ranking = pd.DataFrame({"time":teams, "posicao": np.zeros(len(teams)), "jogos": np.zeros(len(teams)), "pts": np.zeros(len(teams)), "v":np.zeros(len(teams)), "gp":np.zeros(len(teams)), "gc":np.zeros(len(teams)), "sg":np.zeros(len(teams))})
    ranking.set_index('time', inplace=True)
    
    for turn in turns: 
        if turn == 1:
            ranking_DF[season][turn] = classificar(matches.query("ano == " + str(season) + " and rodada == " + str(turn)), turn, ranking)
        else:
            ranking_DF[season][turn] = classificar(matches.query("ano == " + str(season) + " and rodada == " + str(turn)), turn, ranking_DF[season][turn-1])

def getPosicao(classificacao, row, tipo):
    return classificacao[row["ano"]][row["rodada"]].loc[row[tipo]]["posicao"]

matches["posicao_m"] = matches.apply(lambda row: getPosicao(ranking_DF, row, "mandante"),axis=1)

matches["posicao_v"] = matches.apply(lambda row: getPosicao(ranking_DF, row, "visitante"),axis=1)

matches.head(5)

ranking_DF[2018][7]

def aproveitamento5Rodadas (classificacao, row, tipo, jogos):    
    
    rodada = row["rodada"] - 1    
    rodada5r = rodada - jogos
    
    if rodada < 1:
        rodada = 1
        
    if rodada5r < 1:
        rodada5r = 1    
        
    pontos = classificacao[row["ano"]][rodada].loc[row[tipo]]["pts"]
    pontos5r = classificacao[row["ano"]][rodada5r].loc[row[tipo]]["pts"]
    
    if rodada == rodada5r:
        return 0        
    
    return round((pontos - pontos5r)/((rodada - rodada5r)*3), 2)

matches["aproveitamento_5r_v"] = matches.apply(lambda row: aproveitamento5Rodadas(ranking_DF, row, "visitante", 5),axis=1)

matches["aproveitamento_5r_m"] = matches.apply(lambda row: aproveitamento5Rodadas(ranking_DF, row, "mandante", 5),axis=1)

matches["aproveitamento_10r_v"] = matches.apply(lambda row: aproveitamento5Rodadas(ranking_DF, row, "visitante", 10),axis=1)

matches["aproveitamento_10r_m"] = matches.apply(lambda row: aproveitamento5Rodadas(ranking_DF, row, "mandante", 10),axis=1)

matches.head(5)

matches.to_csv(path_or_buf="data/matches-results-positions.9.csv",sep=",")
