==> Step_4-Predictions.ipynb <==
# Campeonato Brasileiro - Série A

## Radom Forest Classifier

### 1. Import packages

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd

from sklearn.ensemble import RandomForestClassifier

from sklearn.ensemble import AdaBoostClassifier

from sklearn.model_selection import cross_val_score

from sklearn.metrics import confusion_matrix
from sklearn.metrics import accuracy_score,precision_score, recall_score
from sklearn.metrics import f1_score

from sklearn import datasets

### 2. Reading data

jogos = pd.read_csv("data/matches-results-positions.9.csv",sep=",")

del jogos["Unnamed: 0"]

#jogos.query('ano > 2012', inplace=True)

jogos.head(3)

jogos.keys()

times = np.concatenate((jogos["visitante"].unique(), jogos["mandante"].unique()), axis=0)

times_dict = dict(zip(times, range(1, len(times))))

times_dict

jogos = jogos.applymap(lambda s: times_dict.get(s) if s in times_dict else s)

### 3. Define target and data subset

jogos_X_df = jogos.copy()

del jogos_X_df["resultado"]

del jogos_X_df["campeonato"]

del jogos_X_df["data"]

del jogos_X_df["gols_mandante"]

del jogos_X_df["gols_visitante"]

del jogos_X_df["aproveitamento_5r_v"]

del jogos_X_df["aproveitamento_5r_m"]

del jogos_X_df["Unnamed: 0.1"]

jogos_X_df.head()

jogos_X_df.keys()

y = jogos["resultado"].values

X = jogos_X_df.values

### 4. Train and Test Split

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(
                                        X,
                                        y, 
                                        test_size=0.2, 
                                        random_state=1)

### 4.1 Grid Search

from sklearn.grid_search import GridSearchCV

from sklearn.datasets import make_classification

from sklearn.ensemble import RandomForestClassifier

rfc = RandomForestClassifier(n_jobs=-1,max_features= 'sqrt' ,n_estimators=50, oob_score = True)

param_grid = {     
    'n_estimators': [800],
    'min_samples_leaf': [10],
    'max_features': ['log2'],
    'n_jobs': [-1],
    'min_samples_split': [3],
    'warm_start': [False],
    'oob_score':  [False], 
    'bootstrap': [True],
    'criterion': ['entropy'],
    'class_weight': ['balanced', 'balanced_subsample', None]
}

CV_rfc = GridSearchCV(estimator=rfc, param_grid=param_grid, cv= 5)

CV_rfc.fit(X, y)

print (CV_rfc.best_params_)

### 5. Apply Random Forest Classifier

rf_model = RandomForestClassifier(random_state=1, criterion='entropy', max_features='log2', n_estimators=800, min_samples_leaf=10)

rf_model.fit(X_train, y_train)

### 6. Model Cross-Validation

cv_scores_rf = cross_val_score(rf_model, X_train, y_train, 
                               cv=5,scoring="accuracy")
cv_scores_rf

cv_scores_rf_mean =  np.mean(cv_scores_rf)
cv_scores_rf_mean

### Apply AdaBoostClassifier

abc_model = AdaBoostClassifier(random_state=1, n_estimators=50, learning_rate=1., algorithm="SAMME.R")

abc_model.fit(X_train, y_train)

cv_scores_abc = cross_val_score(abc_model, X_train, y_train, 
                               cv=5,scoring="accuracy")
cv_scores_abc

cv_scores_abc_mean =  np.mean(cv_scores_abc)
cv_scores_abc_mean

### 7. Model Test

y_test_pred = rf_model.predict(X_test)

y_test.shape

# Results
pd.crosstab(y_test, y_test_pred,
            rownames=['Actual'], 
            colnames=['Predicted'])

confusion_matrix(y_test, y_test_pred)

accuracy_score_test = accuracy_score(y_test, y_test_pred)
accuracy_score_test

#precision_score_forest = precision_score(y_test, y_test_pred)
#precision_score_forest

#recall_score_forest = recall_score(y_test, y_test_pred)
#recall_score_forest

#f1_score_forest = f1_score(y_test, y_test_pred)
#f1_score_forest

### 8. Comparison accuracy: Validation vs Test

comparison = {"Validation": [cv_scores_rf_mean],
             "Test": [accuracy_score_test]}
pd.DataFrame(comparison, index = ["Accuracy"])

### 9. Feature Importance

rf_model.feature_importances_

# View a list of the features and their importance scores
imp_list = list(zip(jogos_X_df.keys(), 
                    rf_model.feature_importances_))
imp_df = pd.DataFrame(imp_list, columns = ["Features", "Importance"])
imp_df.sort_values(by = "Importance", ascending = False)

### 10. Predict

rodada = pd.read_csv("data/turn.10.csv",sep=",")

rodada

rodada_DF = rodada.copy()

rodada_DF = rodada_DF.applymap(lambda s: times_dict.get(s) if s in times_dict else s)

rodada_DF

jogos.head(3)

jogos.query("visitante == 54 and rodada == 30 and ano == 2017")

ano = jogos["ano"].max()

rod = jogos[jogos["ano"] == ano]["rodada"].max()

rodada_DF["ano"] = ano

rodada_DF["rodada"] = rod + 1

def getPosicao(jogos, ano, rodada, time):    

    ultimo_jogo = jogos.query("ano == " + str(ano) + " and rodada == " + str(rodada) + " and mandante == " + str(time))
    if not ultimo_jogo.empty:
        return ultimo_jogo["posicao_m"].values[0]
    
    ultimo_jogo = jogos.query("ano == " + str(ano) + " and rodada == " + str(rodada) + " and visitante == " + str(time))        
    return ultimo_jogo["posicao_v"].values[0]

def getAproveitamento(jogos, ano, rodada, time):    

    ultimo_jogo = jogos.query("ano == " + str(ano) + " and rodada == " + str(rodada) + " and mandante == " + str(time))
    if not ultimo_jogo.empty:
        return ultimo_jogo["aproveitamento_10r_m"].values[0]
    
    ultimo_jogo = jogos.query("ano == " + str(ano) + " and rodada == " + str(rodada) + " and visitante == " + str(time))        
    return ultimo_jogo["aproveitamento_10r_v"].values[0]

jogos

rod

ano

rodada_DF["posicao_m"] = rodada_DF.apply(lambda row: getPosicao(jogos, ano, rod, row["mandante"]),axis=1)

del rodada["placar"]

rodada_DF.head(10)

rodada_DF["posicao_v"] = rodada_DF.apply(lambda row: getPosicao(jogos, ano, rod, row["visitante"]),axis=1)

rodada_DF["aproveitamento_10r_m"] = rodada_DF.apply(lambda row: getAproveitamento(jogos, ano, rod, row["mandante"]),axis=1)

rodada_DF["aproveitamento_10r_v"] = rodada_DF.apply(lambda row: getAproveitamento(jogos, ano, rod, row["visitante"]),axis=1)

rodada_DF

rodada["prediction_RFC"] = rf_model.predict(rodada_DF[["ano", "rodada", "mandante", "visitante", "posicao_m", "posicao_v", "aproveitamento_10r_m", "aproveitamento_10r_v"]].values)

rodada["prediction_ABC"] = abc_model.predict(rodada_DF[["ano", "rodada", "mandante", "visitante", "posicao_m", "posicao_v", "aproveitamento_10r_m", "aproveitamento_10r_v"]].values)

rodada

rodada.to_csv(path_or_buf="data/prediction.10.csv",sep=";")
